requirements.txt
aiofiles==23.2.1
aiohttp==3.9.2
apscheduler==3.6.3
asyncio==3.4.3
beautifulsoup4==4.12.2
bing_image_downloader==1.1.2
bleach==6.1.0
bs4==0.0.1
cachetools==4.2.2
cloudscraper
countryinfo==0.1.2
currencyconverter==0.17.13
dateparser==1.1.8
emoji==2.8.0
emoji-country-flag==1.3.2
feedparser==6.0.10
ffmpeg-python==0.2.0
future==0.18.3
geopy==2.4.0
gpytranslate==1.5.1
hachoir==3.2.0
html2text==2020.1.16
jikanpy==4.3.2
lxml==4.9.3
markdown2==2.4.10
motor==3.3.1
opencv-python-headless==4.8.1.78
pillow==10.2.0
pretty_errors==1.2.25
psutil==5.9.6
psycopg2-binary==2.9.9
pydictionary==2.0.1
pykeyboard==0.1.5
pynewtonmath==1.0.1
pyrate-limiter==3.1.0
pyrogram==2.0.106
python-telegram-bot==13.15
regex==2023.10.3
requests==2.31.0
secureme==0.0.4
search_engine_parser==0.6.8
speedtest-cli==2.1.3
sqlalchemy==1.4.46
telegraph==2.2.0
telethon==1.32.0
tgcrypto==1.2.5
wikipedia==1.4.0
